Refactor Go plugin API and frame builders
